feat: Separate pages and containers for handling dynamic import
