+++ apps/insight-viewer-demo/src/pages/basic/index.tsx
@@ -1,0 +1,8 @@
+import dynamic from 'next/dynamic'
+
+const DynamicComponentWithNoSSR = dynamic(
+  () => import('../../containers/Basic'),
+  { ssr: false }
+)
+
+export default DynamicComponentWithNoSSR
